requirements.txt
requests
mastodon.py